requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==5.0.1
django-debug-toolbar==4.2.0
django-filter==23.5
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pytest==7.4.4
pytils==0.4.1
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
